Fix non-Linux build failure.
gcc-preinclude.h contains a fix for old Linux systems but it's included
unconditionally on other platforms as well.

I could update the GYP files to exclude it on other platforms but it
seems easier to add a `defined(__linux__)` guard to the header file.

Let this be the last fix.