Handling exceptions.
I believe this is the most appropriate way to handle exceptions. Using
v8::Function::Call rather than node::MakeCallback (via NanMakeCallback)
means that these functions will not work with domains, but that seems
appropriate.

I don't know if joyent/node#9245 or nodejs/nan#284 should be a concern
here or not.