Add promise version of database and statement
Add the type PromiseDatabase and PromiseStatement which may be used instead of Database and Statement.
With these versions, the return type of all methods with a callback is converted into a Promise. The user then must not provide the callback, as it is automatically created
for the Promise.
Because constructor cannot return a promise, the Database constructor does not open the database anymore. One has to call PromiseDatabase.open(), which return a callback.
The name of the file for the database has to be given to the constructor while openning mode is expected by the open() method.

Normal Database and Statement can be used at the same time as promise versions, but PromiseDatabase must only be linked to PromiseStatement and vice versa.