Add db.configure('chunkSize', <int>);
There was no way (that i could find) to set the SQLITE_FCNTL_CHUNK_SIZE
via the javascript, so i extended the 'configure' method to allow
'chunkSize' to be passed in.

I also added a callback on the configure method. This commit only has
the change for chunkSize but it probably should be extended to all
configure options (the callback that is).

I created a new ConfigureIntBaton that was very specific to the problem of
configuring the db with a numeric value. I did this instead of
overridding the status value in the baton.  It felt cleaner.

Addresses mapbox/node-sqlite3#980