Static linking option for rebuilding (`--sqlite-static-library`)
Allow users to pass an absolute path to a SQLite static library by
providing a `--sqlite-static-library` flag. This will cause the
binding to be compiled with static linking instead of dynamic
linking. The purpose of allowing this is to allow non-standard
versions of SQLite (i.e. SQLCipher) to be distributed to machines
without the SQLCipher dynamic libraries (i.e. an Electron app that
uses SQLCipher).

Discussion: https://github.com/mapbox/node-sqlite3/issues/1065